Move S3DiscoveryCallback out of callbacks package
